confd: move lyx source to new src lib directory
Preparation for statd.
