+++ src/confd/src/lib/lyx.c
