+++ src/confd/src/lib/lyx.h
